RPG/requirements.txt
streamlit>=1.41.0
pydantic>=2.10.0
PyYAML>=6.0.2
Jinja2>=3.1.5
